Added the informal specification.
Made some changes to the casm file, uploaded it as Skytrain(2).casm.
- Added my name/student# at the top of the file (be sure to add yours!)
- Added an arriving status.
The assignment noted that the sensors can detect when a train
is arriving and stopping.  While the arriving status is essentially useless, I figured the program output might look better if we kept the status in anyway.  I set a trainArriveT for 5 seconds, which will be the period before the status changes from arriving to stopped.